Implement basic array computation for dynamic parallelization
The main goal is to produce ParallelSqrt bytecode from SequentialSqrt automatically.
Currently, SequentialSqrt is faster than ParrallelSqrt. The code has to be profiled to find out why but it is very likely that sqrt() is not complex enough to justify the use of a separate thread with the associated overhead. This does not change anything to the objective though.